requirements.txt
# Core dependencies
python-dotenv>=0.19.0
click>=8.0.0
colorama>=0.4.4
tqdm>=4.62.0

# LLM dependencies
litellm>=1.12.0
langdetect>=1.0.9

# Machine translation
# Uncomment based on which translators you want to use
google-cloud-translate>=3.7.0
# deepl>=1.9.0